refactor: dto명 수정 및 dto 내부 패키지 생성
