+++ src/main/java/com/ryc/api/v1/club/controller/ClubController.java
@@ -1,7 +1,7 @@
 package com.ryc.api.v1.club.controller;
 
-import com.ryc.api.v1.club.dto.CreateClubRequestDto;
-import com.ryc.api.v1.club.dto.CreateClubResponseDto;
+import com.ryc.api.v1.club.dto.request.CreateClubRequest;
+import com.ryc.api.v1.club.dto.response.CreateClubResponse;
 import com.ryc.api.v1.club.service.ClubService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
@@ -18,9 +18,9 @@ public class ClubController {
     private final ClubService clubService;
 
     @PostMapping("/")
-    public ResponseEntity<?> createClub(@Valid @RequestBody CreateClubRequestDto body) {
+    public ResponseEntity<?> createClub(@Valid @RequestBody CreateClubRequest body) {
         try {
-            CreateClubResponseDto responseDto = clubService.createClub(body);
+            CreateClubResponse responseDto = clubService.createClub(body);
             return ResponseEntity.status(HttpStatus.CREATED).body(responseDto);
         } catch (Exception e){
             return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
