+++ src/main/java/com/ryc/api/v1/club/dto/response/CreateClubResponse.java
@@ -1,0 +1,8 @@
+package com.ryc.api.v1.club.dto.response;
+
+import jakarta.validation.constraints.NotEmpty;
+
+import java.time.LocalDateTime;
+
+public record CreateClubResponse(@NotEmpty(message = "created shouldn't be empty") LocalDateTime created) {
+}
